# packages/contracts/openapi.json
{
  "openapi": "3.1.1",
...
  "info": {
    "title": "JSMiner API",
    "version": "0.1.0",
    "description": "Offline content analysis with isolated jsluice and private source artifacts is implemented. URL acquisition, transformations and other extractors remain unavailable."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local development; use a private TLS reverse proxy for remote access"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/analyze": {
      "post": {
        "operationId": "analyze",
        "summary": "Analyze one script",
        "tags": [
          "Analysis"
        ],
        "x-required-permissions": [
          "analysis:write",
          "analysis:read"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response (analysis may be partial or failed)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Permission or capture policy denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Script hash mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Body or script too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported media type or encoding",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid content or semantic parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Capacity or storage quota exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                },
                "description": "Delay before retrying in seconds"
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "URL acquisition is not implemented (offline content only)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Capture failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "Capture timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/source/{handle}": {
      "get": {
        "operationId": "listModules",
        "summary": "List module metadata",
        "tags": [
          "Sources"
        ],
        "x-required-permissions": [
          "source:read"
        ],
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Identifier"
            },
            "description": "Opaque handle bound to the authenticated project"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            },
            "description": "See SourceListQuery"
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "pattern": "^[A-Za-z0-9_-]+$"
            },
            "description": "See SourceListQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response (analysis may be partial or failed)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManifestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Permission or capture policy denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Unknown resource or another project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "Expired handle",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque random token. Project and permissions are resolved server-side; not supplied by the request."
      }
    },
    "schemas": {
      "ToolName": {
        "type": "string",
        "enum": [
          "webcrack",
          "wakaru",
          "jsluice",
          "trufflehog",
          "graphql",
          "domains"
        ],
        "title": "ToolName"
      },
      "Hash": {
        "type": "string",
        "minLength": 71,
        "maxLength": 71,
        "pattern": "^sha256:[0-9a-f]{64}$",
        "title": "Hash"
      },
      "Identifier": {
        "type": "string",
        "minLength": 1,
        "maxLength": 128,
        "pattern": "^[A-Za-z0-9_-]+$",
        "title": "Identifier"
      },
      "HttpUrl": {
        "type": "string",
        "minLength": 1,
        "maxLength": 4096,
        "format": "uri",
        "pattern": "^https?://[^\\s/?#]+(?:[/?#][^\\s]*)?$",
        "title": "HttpUrl"
      },
      "Domain": {
        "type": "string",
        "minLength": 1,
        "maxLength": 253,
        "pattern": "^(?:[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?\\.)+[a-z](?:[a-z0-9-]{0,61}[a-z0-9])?$",
        "title": "Domain"
      },
...
      "AnalyzeRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "url": {
            "$ref": "#/components/schemas/HttpUrl"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10485760
          },
          "tools": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ToolName"
            },
            "minItems": 1,
            "maxItems": 6,
            "uniqueItems": true
          },
          "script_hash": {
            "$ref": "#/components/schemas/Hash"
          },
          "base_url": {
            "$ref": "#/components/schemas/HttpUrl"
          },
          "reference_domains": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Domain"
            },
            "minItems": 0,
            "maxItems": 20,
            "uniqueItems": true
          }
        },
        "required": [],
        "oneOf": [
          {
            "required": [
              "url"
            ]
          },
          {
            "required": [
              "content"
            ]
          }
        ],
        "allOf": [
          {
            "if": {
              "required": [
                "tools"
              ],
              "properties": {
                "tools": {
                  "contains": {
                    "const": "domains"
                  }
                }
              }
            },
            "then": {
              "required": [
                "reference_domains"
              ],
              "properties": {
                "reference_domains": {
                  "minItems": 1
                }
              }
            }
          }
        ],
        "description": "Exactly one of url/content. content must also be well-formed Unicode and <= 10 MiB UTF-8 (runtime refinement). Explicit domains requires reference_domains.",
        "title": "AnalyzeRequest"
      },
...
      "Confidence": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "Confidence"
      },
...
      "Endpoint": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Identifier"
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence"
          },
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "minItems": 1,
            "maxItems": 5
          },
          "value": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2048
          },
          "resolved_url": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HttpUrl"
              },
              {
                "type": "null"
              }
            ]
          },
          "method": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 32,
                "pattern": "^[A-Z]+$"
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "type": "string",
            "enum": [
              "http_call",
              "navigation",
              "resource",
              "literal"
            ]
          },
          "dynamic": {
            "type": "boolean"
          },
          "query_params": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "minItems": 0,
            "maxItems": 128,
            "uniqueItems": true
          },
          "body_params": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "minItems": 0,
            "maxItems": 128,
            "uniqueItems": true
          }
        },
        "required": [
          "id",
          "confidence",
          "evidence",
          "value",
          "resolved_url",
          "method",
          "kind",
          "dynamic",
          "query_params",
          "body_params"
        ],
        "title": "Endpoint"
      },
      "Secret": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Identifier"
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence"
          },
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SecretEvidence"
            },
            "minItems": 1,
            "maxItems": 5
          },
          "kind": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "rule_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "pattern": "^(?:jsluice|trufflehog):[A-Za-z0-9._-]+$"
          },
          "masked_value": {
            "const": "[REDACTED]",
            "type": "string"
          },
          "fingerprint": {
            "type": "string",
            "minLength": 76,
            "maxLength": 76,
            "pattern": "^hmac-sha256:[0-9a-f]{64}$"
          },
          "validation": {
            "type": "string",
            "const": "not_performed"
          }
        },
        "required": [
          "id",
          "confidence",
          "evidence",
          "kind",
          "rule_id",
          "masked_value",
          "fingerprint",
          "validation"
        ],
        "title": "Secret"
      },
      "GraphqlOperation": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "$ref": "#/components/schemas/Identifier"
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence"
          },
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "minItems": 1,
            "maxItems": 5
          },
          "operation_type": {
            "type": "string",
            "enum": [
              "query",
              "mutation",
              "subscription"
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 128,
                "pattern": "^[_A-Za-z][_0-9A-Za-z]*$"
              },
              {
                "type": "null"
              }
            ]
          },
          "variables": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "properties": {
                "name": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                "type": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 256
                }
              },
              "required": [
                "name",
                "type"
              ]
            },
            "minItems": 0,
            "maxItems": 128
          },
          "root_fields": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "minItems": 0,
            "maxItems": 128,
            "uniqueItems": true
          },
          "document_hash": {
            "$ref": "#/components/schemas/Hash"
          },
          "endpoint_id": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Identifier"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "confidence",
          "evidence",
          "operation_type",
          "name",
          "variables",
          "root_fields",
          "document_hash",
          "endpoint_id"
        ],
        "title": "GraphqlOperation"
      },
...
      "CoverageStatus": {
        "type": "string",
        "enum": [
          "complete",
          "partial",
          "failed",
          "not_requested"
        ],
        "title": "CoverageStatus"
      },
...
      "AnalyzeResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "0.1"
          },
          "handle": {
            "$ref": "#/components/schemas/Identifier"
          },
          "status": {
            "type": "string",
            "enum": [
              "complete",
              "partial",
              "failed"
            ]
          },
          "script_hash": {
            "$ref": "#/components/schemas/Hash"
          },
          "expires_at": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40,
            "format": "date-time"
          },
          "cache": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "hit",
                  "partial_hit",
                  "miss"
                ]
              }
            },
            "required": [
              "status"
            ]
          },
          "endpoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Endpoint"
            },
            "minItems": 0,
            "maxItems": 200
          },
          "secrets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Secret"
            },
            "minItems": 0,
            "maxItems": 200
          },
          "gql_operations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GraphqlOperation"
            },
            "minItems": 0,
            "maxItems": 200
          },
          "subdomains": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Subdomain"
            },
            "minItems": 0,
            "maxItems": 200
          },
          "coverage": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "endpoints": {
                "$ref": "#/components/schemas/CoverageStatus"
              },
              "secrets": {
                "$ref": "#/components/schemas/CoverageStatus"
              },
              "gql_operations": {
                "$ref": "#/components/schemas/CoverageStatus"
              },
              "subdomains": {
                "$ref": "#/components/schemas/CoverageStatus"
              }
            },
            "required": [
              "endpoints",
              "secrets",
              "gql_operations",
              "subdomains"
            ]
          },
          "tools": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ToolRun"
            },
            "minItems": 1,
            "maxItems": 6
          },
          "truncation": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "truncated": {
                "type": "boolean"
              },
              "reasons": {
                "type": "array",
                "items": {
                  "type": "string",
                  "enum": [
                    "response_bytes",
                    "finding_count",
                    "evidence_count",
                    "artifact_bytes",
                    "module_count",
                    "field_bytes"
                  ]
                },
                "minItems": 0,
                "maxItems": 6,
                "uniqueItems": true
              }
            },
            "required": [
              "truncated",
              "reasons"
            ]
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "properties": {
                "code": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128,
                  "pattern": "^[a-z][a-z0-9_]*$"
                },
                "tool": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ToolName"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              },
              "required": [
                "code",
                "tool"
              ]
            },
            "minItems": 0,
            "maxItems": 32
          }
        },
        "required": [
          "schema_version",
          "handle",
          "status",
          "script_hash",
          "expires_at",
          "cache",
          "endpoints",
          "secrets",
          "gql_operations",
          "subdomains",
          "coverage",
          "tools",
          "truncation",
          "warnings"
        ],
        "title": "AnalyzeResponse"
      },
...
      "ManifestResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "handle": {
            "$ref": "#/components/schemas/Identifier"
          },
          "expires_at": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40,
            "format": "date-time"
          },
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            },
            "minItems": 0,
            "maxItems": 100
          },
          "total_modules": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2000
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 512,
                "pattern": "^[A-Za-z0-9_-]+$"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "handle",
          "expires_at",
          "modules",
          "total_modules",
          "next_cursor"
        ],
        "title": "ManifestResponse"
      },
...
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1,
                "maxLength": 128,
                "pattern": "^[a-z][a-z0-9_]*$"
              },
              "message": {
                "type": "string",
                "minLength": 1,
                "maxLength": 512
              },
              "request_id": {
                "$ref": "#/components/schemas/Identifier"
              }
            },
            "required": [
              "code",
              "message",
              "request_id"
            ]
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
    }
  }
}